src/requirements.txt
torch>=1.9.0
torchvision>=0.10.0
opencv-python>=4.5.0
pillow>=8.0.0
numpy>=1.21.0
pandas>=1.3.0
tqdm>=4.62.0
requests>=2.25.0
pycocotools>=2.0.2
pushbullet.py>=0.12.0
flask>=2.0.0
flask-cors>=3.0.0